requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.61.0
anyio==4.10.0
asyncer==0.0.8
attrs==25.3.0
backoff==2.2.1
bidict==0.23.1
certifi==2025.8.3
chainlit==2.6.7
charset-normalizer==3.4.2
chevron==0.14.0
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
eval_type_backport==0.2.2
fastapi==0.116.1
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.7.0
googleapis-common-protos==1.70.0
griffe==1.11.0
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
Lazify==0.4.0
literalai==0.1.201
logfire-api==4.2.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.12.4
monotonic==1.6
multidict==6.6.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
openai==1.99.3
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-alephalpha==0.44.1
opentelemetry-instrumentation-anthropic==0.44.1
opentelemetry-instrumentation-bedrock==0.44.1
opentelemetry-instrumentation-chromadb==0.44.1
opentelemetry-instrumentation-cohere==0.44.1
opentelemetry-instrumentation-crewai==0.44.1
opentelemetry-instrumentation-google-generativeai==0.44.1
opentelemetry-instrumentation-groq==0.44.1
opentelemetry-instrumentation-haystack==0.44.1
opentelemetry-instrumentation-lancedb==0.44.1
opentelemetry-instrumentation-langchain==0.44.1
opentelemetry-instrumentation-llamaindex==0.44.1
opentelemetry-instrumentation-logging==0.55b1
opentelemetry-instrumentation-marqo==0.44.1
opentelemetry-instrumentation-mcp==0.44.1
opentelemetry-instrumentation-milvus==0.44.1
opentelemetry-instrumentation-mistralai==0.44.1
opentelemetry-instrumentation-ollama==0.44.1
opentelemetry-instrumentation-openai==0.44.1
opentelemetry-instrumentation-openai-agents==0.44.1
opentelemetry-instrumentation-pinecone==0.44.1
opentelemetry-instrumentation-qdrant==0.44.1
opentelemetry-instrumentation-redis==0.55b1
opentelemetry-instrumentation-replicate==0.44.1
opentelemetry-instrumentation-requests==0.55b1
opentelemetry-instrumentation-sagemaker==0.44.1
opentelemetry-instrumentation-sqlalchemy==0.55b1
opentelemetry-instrumentation-threading==0.55b1
opentelemetry-instrumentation-together==0.44.1
opentelemetry-instrumentation-transformers==0.44.1
opentelemetry-instrumentation-urllib3==0.55b1
opentelemetry-instrumentation-vertexai==0.44.1
opentelemetry-instrumentation-watsonx==0.44.1
opentelemetry-instrumentation-weaviate==0.44.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-semantic-conventions-ai==0.4.11
opentelemetry-util-http==0.55b1
packaging==25.0
posthog==3.25.0
propcache==0.3.2
protobuf==5.29.5
pydantic==2.11.7
pydantic-ai-slim==0.6.2
pydantic-graph==0.6.2
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.18
python-socketio==5.13.0

PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.47.2
syncer==2.0.3
tenacity==9.1.2
tokenizers==0.21.4
tomli==2.2.1
tqdm==4.67.1
traceloop-sdk==0.44.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==0.20.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
